+++ g2p_registrant/models/__init__.py
@@ -15,6 +15,7 @@
 from . import registrant
 from . import individual
 from . import group
+from . import registrant_attribute
 from . import reg_id
 from . import reg_relationship
 from . import group_membership